Created a API based Advice Generator
This is one of my first PRs and I want to contribute in this field and thus I want to get to know more about Open-source and contribute more and hence want this first Fest of mine to be a successful one. As most of them out there are excluded projects and hence I would request you to give me a chance to create a second PR in your project. I have come up with another project that is an API based Advice Generator I hope you will find it nice and help me to get my 2nd PR in in this fest